requirements.txt
osu.py>=1.2.0,<2
requests>=2.25.1,<3
uvicorn>=0.23.2,<1
git+https://github.com/Sheepposu/osu_diff_calc@423b06a5e770ce90467ad1d9da0b8e95dc260382#egg=osu_diff_calc